Fix conflicts
merged ca_analysis
